refactor: 암복호화 로직 위치변경
